Restructure storybook
