+++ .storybook/stories/components/tagContextMenu/TagContextMenuWrapper.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
-import TagContextMenu from '../../../src/components/tagContextMenu/TagContextMenu';
-import TagIcon from '../../../src/elements/tagIcon/TagIcon';
-import { getIcon } from '../../../src/helpers/getIcon';
+import TagContextMenu from '../../../../src/components/tagContextMenu/TagContextMenu';
+import TagIcon from '../../../../src/elements/icons/tagIcon/TagIcon';
+import { getIcon } from '../../../../src/helpers/getIcon';
 
 export interface TagContextMenuWrapperProps {
     icon: string;
