+++ .storybook/stories/components/tagInfoButton/TagMoreInfo.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta, Story } from '@storybook/react/types-6-0';
 import React from 'react';
-import DataInfoButton, { DataInfoButtonProps } from '../../../src/components/contextMenuPopover/DataInfoButton';
+import DataInfoButton, { DataInfoButtonProps } from '../../../../src/components/contextMenuPopover/DataInfoButton';
 
 export default {
     title: 'TagInfoButton',
