+++ .storybook/stories/structure/iconList/IconListWrapper.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { IconList, IconListProps } from '../../../src/structure/iconList/IconList';
+import { IconList, IconListProps } from '../../../../src/structure/iconList/IconList';
 
 const IconListWrapper: React.FC<IconListProps> = ({ style, title, items, isMovable, expanded }: IconListProps) => {
     return (
