+++ .storybook/stories/structure/linkCard/LinkCardWrapper.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { LinkCard, LinkCardProps } from '../../../src/structure/linkCard/LinkCard';
+import { LinkCard, LinkCardProps } from '../../../../src/structure/linkCard/LinkCard';
 
 const LinkCardWrapper: React.FC<LinkCardProps> = ({ style, image, title, subTitle, links }: LinkCardProps) => {
     return (
